+++ src/main/java/gift/auth/service/KakaoService.java
@@ -1,91 +1,62 @@
 package gift.auth.service;
 
-
+import gift.auth.client.KakaoOauthClient;
 import gift.auth.dto.KakaoTokenResponseDto;
 import gift.auth.dto.KakaoUserInfoDto;
 import gift.exception.ErrorCode;
-import gift.exception.KakaoApiErrorException;
-import gift.exception.KakaoClientErrorException;
 import gift.exception.KakaoLoginErrorException;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatusCode;
-import org.springframework.http.MediaType;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import gift.user.JwtTokenProvider;
+import gift.user.entity.User;
+import gift.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestClient;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@Transactional(readOnly = true)
 public class KakaoService {
 
-  private final String clientId;
-  private final String redirectUrl;
+  private final KakaoOauthClient kakaoOauthClient;
+  private final UserRepository userRepository;
+  private final JwtTokenProvider jwtTokenProvider;
 
-  private static final String KAKAO_AUTH_URL = "https://kauth.kakao.com";
-  private static final String KAKAO_TOKEN_URL = "https://kauth.kakao.com/oauth/token";
-
-  private final RestClient restClient;
-
-  public KakaoService(@Value("${kakao.client.id}") String clientId,
-      @Value("${kakao.redirect.url}") String redirectUrl) {
-    this.clientId = clientId;
-    this.redirectUrl = redirectUrl;
-
-    HttpComponentsClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory();
-
-    requestFactory.setConnectTimeout(5000);
-    requestFactory.setReadTimeout(5000);
-    requestFactory.setConnectionRequestTimeout(5000);
-
-    this.restClient = RestClient.builder()
-        .requestFactory(requestFactory)
-        .build();
+  public KakaoService(KakaoOauthClient kakaoOauthClient,
+      UserRepository userRepository,
+      JwtTokenProvider jwtTokenProvider) {
+    this.kakaoOauthClient = kakaoOauthClient;
+    this.userRepository = userRepository;
+    this.jwtTokenProvider = jwtTokenProvider;
   }
 
-  // Step 1: 카카오 로그인 URL 생성 사용자를 카카오 인증 서버로 리다이렉트하기 위한 URL
   public String getKakaoLoginUrl() {
-    return KAKAO_AUTH_URL + "/oauth/authorize" +
-        "?scope=talk_message" +
-        "&response_type=code" +
-        "&redirect_uri=" + redirectUrl +
-        "&client_id=" + clientId;
+    return kakaoOauthClient.getKakaoLoginUrl();
   }
 
-  //Step2: 인가 코드로 액세스 토큰 받기
+  @Transactional
+  public String processKakaoLogin(String authorizationCode) {
+    try {
+      KakaoTokenResponseDto tokenResponse = kakaoOauthClient.getAccessToken(authorizationCode);
+      KakaoUserInfoDto userInfo = kakaoOauthClient.getUserInfo(tokenResponse.accessToken());
+      User user = findOrCreateUser(userInfo);
+      return jwtTokenProvider.generateToken(user);
 
-  public KakaoTokenResponseDto getAccessToken(String authorizationCode) {
-    MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-    body.add("grant_type", "authorization_code");
-    body.add("client_id", clientId);
-    body.add("redirect_uri", redirectUrl);
-    body.add("code", authorizationCode);
-
-    return restClient.post()
-        .uri(KAKAO_TOKEN_URL)
-        .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-        .body(body)
-        .retrieve()
-        .onStatus(HttpStatusCode::is4xxClientError, (request, responseEntity) -> {
-          throw new KakaoClientErrorException();
-        })
-        .onStatus(HttpStatusCode::is5xxServerError, (request, responseEntity) -> {
-          throw new KakaoApiErrorException(ErrorCode.KAKAO_API_ERROR);
-        })
-        .body(KakaoTokenResponseDto.class);
+    } catch (Exception e) {
+      throw new KakaoLoginErrorException(ErrorCode.KAKAO_LOGIN_ERROR);
+    }
   }
 
-  public KakaoUserInfoDto getUserInfo(String accessToken) {
-    return restClient.get()
-        .uri("https://kapi.kakao.com/v2/user/me")
-        .header("Authorization", "Bearer " + accessToken)
-        .retrieve()
-        .onStatus(HttpStatusCode::is4xxClientError, (request, response) -> {
-          throw new KakaoClientErrorException();
-        })
-        .onStatus(HttpStatusCode::is5xxServerError, (request, response) -> {
-          throw new KakaoLoginErrorException(ErrorCode.KAKAO_LOGIN_ERROR);
-        })
-        .body(KakaoUserInfoDto.class);
+  private User findOrCreateUser(KakaoUserInfoDto userInfo) {
+    String email = userInfo.getEmailSafely();
+    String kakaoId = userInfo.id().toString();
+
+    return userRepository.findByEmail(email)
+        .orElseGet(() -> createNewKakaoUser(email, kakaoId));
+  }
+
+
+  private User createNewKakaoUser(String email, String kakaoId) {
+    String dummyPassword = "KAKAO_LOGIN_" + kakaoId;
+    User newUser = new User(email, dummyPassword);
+
+    return userRepository.save(newUser);
   }
 }
